requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18       
aiosignal==1.3.2       
ansicon==1.89.0        
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blessed==1.21.0
branca==0.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cramjam==2.10.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.2.0
Faker==37.1.0
fastjsonschema==2.21.1
fastparquet==2024.11.0
filelock==3.18.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
gdown==5.2.0
grandalf==0.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
igraph==0.11.8
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipyleaflet==0.19.2
ipython==8.18.1
ipywidgets==8.1.6
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-leaflet==0.19.2
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
keyboard==0.13.5
kiwisolver==1.4.7
labelImg==1.8.6
lap==0.5.12
libusb==1.0.27.post4
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==3.1.3
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.0.0
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netgraph==4.13.2
networkx==3.2.1
notebook==7.4.2
notebook_shim==0.2.4
numpy==2.0.2
opencv-python==4.11.0.86
openvino==2025.1.0
openvino-telemetry==2025.1.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
pkg_about==1.2.11
platformdirs==4.3.7
progressbar==2.5
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py2cytoscape==0.7.1
pyarrow==20.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydot==4.0.0
pydotplus==2.0.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pynput==1.8.1
pyparsing==3.2.3
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-escpos==3.1
python-igraph==0.11.8
python-json-logger==3.3.0
python-magic-bin==0.4.14
pytweening==1.2.0
pytz==2025.2
pyusb==1.3.1
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
qrcode==8.2
rectangle-packer==2.0.4
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
simple-image-download==0.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
texttable==1.7.0
tinycss2==1.4.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.105
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
xyzservices==2025.1.0
yarl==1.20.0
zipp==3.21.0